safari-600.7-branch/LayoutTests/sputnik/Conformance/12_Statement/12.5_The_if_Statement/S12.5_A8-expected.txt

CONSOLE MESSAGE: line 78: SyntaxError: Unexpected token ')'
S12.5_A8

PASS Expected parsing failure

TEST COMPLETE
